SIMSBIOHUB-192: PR feedback 2
